+++ .gitignore
@@ -1,3 +1,6 @@
 *.iml
 .idea/
+.gradle/
 target/
+build
+bukkit/build
