+++ build.gradle.kts
@@ -1,0 +1,14 @@
+buildscript {
+    dependencies {
+        classpath("ru.astrainteractive.gradleplugin:convention:0.5.1")
+    }
+}
+
+plugins {
+    java
+    `java-library`
+    alias(libs.plugins.kotlin.jvm) apply false
+    alias(libs.plugins.klibs.gradle.java.core) apply false
+    alias(libs.plugins.klibs.gradle.rootinfo)
+    alias(libs.plugins.gradle.shadow) apply false
+}
