+++ bukkit/src/main/java/ru/overwrite/protect/bukkit/PasswordHandler.java
