+++ bukkit/src/main/java/ru/overwrite/protect/bukkit/api/ServerProtectorPasswordSuccessEvent.java
