+++ bukkit/src/main/java/ru/overwrite/protect/bukkit/utils/Config.java
