+++ bukkit/src/main/java/ru/overwrite/protect/bukkit/utils/logging/PaperLogger.java
