+++ bukkit/src/main/resources-ru/config.yml
