+++ bukkit/src/main/resources-ru/plugin.yml
@@ -1,5 +1,5 @@
-name: UltimateServerProtector
-version: 26.0
+name: '${name}'
+version: '${version}'
 author: OverwriteMC
 main: ru.overwrite.protect.bukkit.ServerProtector
 api-version: 1.13
