+++ gradle.properties
@@ -1,0 +1,12 @@
+kotlin.code.style=official
+# Java
+makeevrserg.java.source=8
+makeevrserg.java.target=17
+makeevrserg.java.ktarget=17
+# Project
+makeevrserg.project.name=UltimateServerProtector
+makeevrserg.project.group=ru.overwrite.protect.bukkit
+makeevrserg.project.version.string=26.0
+makeevrserg.project.description=Incredibly lightweight plugin, that add an "admin-password" to your server
+makeevrserg.project.developers=Overwrite987|Overwrite987|Overwrite987
+makeevrserg.project.url=https://github.com/Overwrite987/UltimateServerProtector
