+++ gradle/libs.versions.toml
@@ -1,0 +1,17 @@
+[versions]
+# klibs
+klibs-gradleplugin = "0.5.0"
+
+[libraries]
+# Minecraft
+minecraft-bstats = { module = "org.bstats:bstats-bukkit", version.strictly = "3.0.0" }
+minecraft-folia = { module = "dev.folia:folia-api", version.strictly = "1.20.1-R0.1-SNAPSHOT" }
+
+[plugins]
+# Kotlin
+kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.strictly = "1.9.0" }
+
+# Gradle
+gradle-shadow = { id = "com.github.johnrengelman.shadow", version.strictly = "7.1.2" }
+klibs-gradle-java-core = { id = "ru.astrainteractive.gradleplugin.java.core", version.ref = "klibs-gradleplugin" }
+klibs-gradle-rootinfo = { id = "ru.astrainteractive.gradleplugin.root.info", version.ref = "klibs-gradleplugin" }
