+++ settings.gradle.kts
@@ -1,0 +1,44 @@
+pluginManagement {
+    repositories {
+        mavenLocal()
+        mavenCentral()
+        gradlePluginPortal()
+        maven("https://repo.extendedclip.com/content/repositories/placeholderapi/")
+        maven("https://hub.spigotmc.org/nexus/content/repositories/snapshots/")
+        maven("https://papermc.io/repo/repository/maven-public/")
+        maven("https://repo.papermc.io/repository/maven-public/")
+        maven("https://nexus.scarsz.me/content/groups/public/")
+        maven("https://repo.dmulloy2.net/repository/public/")
+        maven("https://oss.sonatype.org/content/groups/public/")
+        maven("https://repo.essentialsx.net/snapshots/")
+        maven("https://repo.maven.apache.org/maven2/")
+        maven("https://maven.enginehub.org/repo/")
+        maven("https://repo1.maven.org/maven2/")
+        maven("https://maven.playpro.com")
+        maven("https://jitpack.io")
+    }
+}
+dependencyResolutionManagement {
+    repositories {
+        mavenLocal()
+        mavenCentral()
+        maven("https://repo.extendedclip.com/content/repositories/placeholderapi/")
+        maven("https://hub.spigotmc.org/nexus/content/repositories/snapshots/")
+        maven("https://papermc.io/repo/repository/maven-public/")
+        maven("https://repo.papermc.io/repository/maven-public/")
+        maven("https://nexus.scarsz.me/content/groups/public/")
+        maven("https://repo.dmulloy2.net/repository/public/")
+        maven("https://oss.sonatype.org/content/groups/public/")
+        maven("https://repo.essentialsx.net/snapshots/")
+        maven("https://repo.maven.apache.org/maven2/")
+        maven("https://maven.enginehub.org/repo/")
+        maven("https://m2.dv8tion.net/releases")
+        maven("https://repo1.maven.org/maven2/")
+        maven("https://maven.playpro.com")
+        maven("https://jitpack.io")
+    }
+}
+enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
+rootProject.name = "UltimateServerProtector"
+
+include(":bukkit")
